+++ lib/commands/slash/anime.py
@@ -1,5 +1,5 @@
 from settings import *
-from load import *
+from objects import *
 
 
 json = {
